refactor(agents): extract run orchestration helpers (#72072)
(cherry picked from commit 29617d543ad07b87d1cbd3800db89345ef622e16)
